refactor(ui): rehome button primitive
