+++ npm/ui/core/src/families/actions/button/button.ts
@@ -1,0 +1,4 @@
+export * from "./button-keyboard.ts";
+
+/** Accessible, unstyled button with polymorphic rendering. */
+export { default as Button } from "./button.vue";
